desktop/requirements.txt
pywebview==5.3
pyinstaller==6.11.1
